test 5

+++ src/controller/uploadCtrl.js
@@ -1,22 +1,28 @@
 const asyncHandler = require('express-async-handler');
-
 const { cloudinaryUploadImg, cloudinaryDeleteImg } = require('../untils/cloudinary');
 const fs = require('fs');
+const { productImagResize } = require('../middlewares/uploadImages');
 
 ////upload image
 
 const uploadImages = asyncHandler(async (req, res) => {
     try {
-        const uploader = (path) => cloudinaryUploadImg(path, 'images');
-        const images = [];
-        const files = req.files;
-        for (const file of files) {
-            const { path } = file;
-            const { url, asset_id, public_id } = await uploader(path);
-            images.push({ url, asset_id, public_id });
-            fs.unlinkSync(path);
-        }
-        res.json(images);
+        await productImagResize(req, res, async () => {
+            const files = req.files;
+            if (!files || files.length === 0) {
+                throw new Error('no file loads');
+            }
+            const images = [];
+            await Promise.all(
+                files.map(async (file) => {
+                    const { path } = file;
+                    const uploadedImage = await cloudinaryUploadImg(path);
+                    images.push(uploadedImage);
+                    fs.unlinkSync(path);
+                }),
+            );
+            res.json(images);
+        });
     } catch (err) {
         throw new Error(err);
     }
